+++ CHANGELOG.md
@@ -14,6 +14,13 @@
 
 - nothing removed
 
+## [1.35.0] - 2021-10-22
+
+### Added
+
+- Logs for auth views (Register/Change Password/Secure Login)
+- last_login update for auth views
+
 ## [1.34.0] - 2021-09-03
 
 ### Added